kallsyms: skip user-space mapped addresses
RHEL9 s390x with KPTI enabled has been observed to sporadically
use same address ranges for user-space libraries as the ones
occupied by kallsyms while in kernel-space.

The test currently (wrongly) assumes, that a write to kallsym
address should fail and doesn't take into account that same
memory ranges could be mapped differently in user vs. kernel-space.

Add a check to see if a kallsym address is already mapped by user-space,
and if it is, skip it to avoid false positives.

Before:
    13	kallsyms.c:115: TFAIL: Successfully accessed kernel addr 0x3ff8060787d (r) (__kstrtabns_nfs3_set_ds_client)
    14	kallsyms.c:115: TFAIL: Successfully accessed kernel addr 0x3ff806007a0 (t) (nfs_init_server_aclclient)
    15	kallsyms.c:115: TFAIL: Successfully accessed kernel addr 0x3ff8060a058 (d) (nfsacl_version)
    16	kallsyms.c:115: TFAIL: Successfully accessed kernel addr 0x3ff8060a078 (d) (nfsacl_rpcstat)
    ...
    kallsyms.c:130: TFAIL: Caught 103102 SIGSEGV but expected 153738

After:
    kallsyms.c:175: TDEBUG: Skipping userspace mapped address 0x3ff80220024
    kallsyms.c:175: TDEBUG: Skipping userspace mapped address 0x3ff8022003c
    kallsyms.c:175: TDEBUG: Skipping userspace mapped address 0x3ff80213b20
    ...
    kallsyms.c:184: TPASS: Caught 103473 SIGSEGV in access ksymbols addr, skipped 27234

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>