tmpdir: rename tst_tmpdir_mkpath to tst_tmpdir_genpath
When I reviewed the rename15 patch I felt confused a while on
function like tst_tmpdir_mkpath (reminds me of: mkdir).

Because the name could be misleading since it suggests that a
file or directory is being created, when in fact it is simply
constructing a path inside a temporary directory without
actually creating any files.

To make the function's purpose clearer, the name should reflect
that it is only constructing or generating a path, not creating
any files or directories.

So I think either 'tst_tmpdir_genpath' or 'tst_tmpdir_buildpath'
would be concise and clear options.

Signed-off-by: Li Wang <liwang@redhat.com>
Acked-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>