lib/tst_test.c: Make use of the run_tcase_on_fs() return value
When the support for per-testcase FS parameters was added the
code that prepares a device along with the fork_testrun() call was put
into a run_tcase_on_fs() function. However by a mistake the return value
from this function wasn't used properly.

The difference is that the test library did exit the whole test if a
single filesystem iteration reported TBROK or an invalid return value.

Without this patch the library carries on even if there was a fatal
problem reported.

Fixes: cce6188916b9 ('lib: tst_test: Add per filesystem mkfs and mount opts')
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Andrea Cervesato <andrea.cervesato@suse.com>