Revert "syscalls: Add missing exit status check"
This reverts commit ffdd3b36cd90ad3f4e199344d20c00fb06762eb2.

This brought a regression to select03.c on 64bit which does not support
__NR__newselect:

select_var.h:89: TCONF: syscall(-1) __NR__newselect not supported on your arch
select03.c:90: TFAIL: Child exited with 32

This is due code in tst_vbrk_() which exits with the test with a return
value in the case of a child processes. Once this is fixed, this change
can be applied again.

Link: https://lore.kernel.org/ltp/ZzSr-X47F4MfM831@yuki.lan/
Acked-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>