process_madvise01: running the test in mem_cg
The MADV_PAGEOUT behavior in the kernel is advisory and may skip
swapping if the system has sufficient free RAM, even when the
advice is explicitly requested. This causes sporadic false positives
in our CI, particularly on systems with large amounts of RAM:

  process_madvise01.c:38: TINFO: Allocate memory: 1048576 bytes
  process_madvise01.c:99: TINFO: Reclaim memory using MADV_PAGEOUT
  process_madvise01.c:62: TFAIL: Expect: Most of the memory has been swapped out: 0kB out of 1024kB

To address this, the patch confines the test to a memory cgroup
with configured limits for memory.max and memory.swap.max, improving
control over memory and swap usage. This reduces the likelihood of
false positives caused by system-wide memory conditions.

Signed-off-by: Li Wang <liwang@redhat.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>