cpuset_memory_spread_test:Modify the creation of fifo files
When the cgroup file system cannot be mounted,
the test will be skipped directly in the setup() of cpuset_funcs.sh.
This will cause the created myfifo file to not be deleted.
When the use case is run again,
an error will be reported that myfifo already exists.

Therefore, removing the $FIFO before we attempt to create.

Signed-off-by: Deng Yu 1471640443dy@gmail.com