lib/tst_res_.c: Add TBROK handler + more verbose errors
We use the tst_res_ helper for tst_brk_ as well so we need to be able to
handle TBROK type as well.

Note that we pass TBROK to the tst_res() here deliberately here because
we want to return back to the shell library, which is the process that
reports the TBROK to the test library by doing 'exit 2'. This works
because we pass the value by a variable as compiler cannot evaluate it's
value at a compile time.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>