lib: PAIR_LOOKUP(): Print int value instead of "???"
When there was no mapping found the PAIR_LOOKUP() returned "???" which
wasn't really helpful. This commit changes the macro to print the integer
value and return a pointer to a static buffer with the value instead.

This means that functions such as tst_strsig() and tst_strerrno() will
now print a numerical value instead of "???" when there is no mapping
defined.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>