lib: Remove CONFIG_LATENCYTOP as "slow" kernel option
CONFIG_LATENCYTOP enables a kernel latency tracking infrastructure that is
used by the "latencytop" userspace tool. The latency that is tracked is not
else consuming CPU), but instead, it is the latency an application encounters
because the kernel sleeps on its behalf for various reasons.

This code tracks 2 levels of statistics:
1) System level latency
2) Per process latency

IMHO this option is safe unless one runs latencytop userspace tool while
testing LTP.

Motivation for this is the openSUSE and some SLES kernels use
CONFIG_LATENCYTOP and all tests are working with a default timeout.
starvation.c is disabled with TCONF on slow kernels, which previously
worked well on kernels with CONFIG_LATENCYTOP.

Link: https://lore.kernel.org/ltp/20250120143420.815363-1-pvorel@suse.cz/
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>