syscalls/mallinfo02: introduce LTP_VAR_USED to avoid optimization
gcc 15 is a bit more clever and noticed that 'buf' isn't used
for anything so it optimized it out, including call to malloc.
So, there's also no mmap() call behind it and test fails,
because nothing was allocated.

Since __attribute__((used)) is ignored for local variables,
introduce LTP_VAR_USED macro, that makes compiler aware of the
variable and doesn't optimize it out.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>