fanotify24: add mmap() and user page fault to test
Commit 066e053fe208a ("fsnotify: add pre-content hooks on mmap()")
in 6.14-rc7 added a pre-content hooh in mmap() before the final release
of the fsnotify pre-content event feature.

To test pre-content hook on mmap(), increase the size of the test file
to 101 pages use mmap() to setup a buffer from the end of the test file.

Change some of the test cases to allow the write() and use this buffer
as the input buffer to write(), which tests deadlock avoidance by
suppresing pre-content hook on user page faults.

Link: https://lore.kernel.org/ltp/20250319192742.999506-4-amir73il@gmail.com/
Reviewed-by: Jan Kara <jack@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>