shell_loader_filesystems: add tmpfs support to avoid fail in GitHub CI
GitHub CI currently fails when running shell_loader_filesystems.sh because
the Docker environment does not have support for XFS and Btrfs. This results
in a TCONF failure, causing the CI job to fail:

  * shell_loader_filesystems.sh (TCONF, exp: TPASS)
  make: *** [Makefile:210: test-shell-loader] Error 1

To resolve this, we add tmpfs to the list of tested filesystems. Since tmpfs
is always available in GitHub CI, this ensures that the test runs successfully
even when XFS and Btrfs are unavailable.

CI: https://github.com/wangli5665/ltp/actions/runs/13992429178

Signed-off-by: Li Wang <liwang@redhat.com>