memcontrol03: Account for process size in cgroup allocation
The first trunk_G allocation has 2MB safety margin to avoid triggering
OOM killer. However, on systems with 64K pagesize, this may not be
enough.  Account for existing cgroup memory usage before, and possible
overcharge, as reported by cgroup memory stats before allocating memory
in child processes.

There was an attempt to fix this in kernel but the problem was concluded
to be artificial [1].

[1] https://lore.kernel.org/lkml/20250530151858.672391-1-mkoutny@suse.com/

Signed-off-by: Martin Doucha <mdoucha@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>