overcommit_memory: Disable optimization for malloc to prevent false positives
When compiling with GCC 15.1.1+, the default-enabled -fmalloc-dce=2 and
-fallocation-dce under -O2 can eliminate the entire malloc block because its
result is only compared to NULL and/or passed to free().

This leads to false positives in tests that expect malloc() to fail under memory
pressure, as the allocation is optimized away.

Disable this optimization for the affected function to preserve the intended
test behavior.

References: https://gcc.gnu.org/onlinedocs/gcc-15.1.0/gcc/Optimize-Options.html
Signed-off-by: Jan Polensky <japo@linux.ibm.com>
Suggested-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>