syscalls: kill: Remove kill07
The test case description reads:

"Test case to check that SIGKILL can not be caught."

It makes use of sigaction() to set up a handler for SIGKILL, but
according to its manual page:

"EINVAL An invalid signal was specified.  This will also be generated if an
attempt is made to change the action for SIGKILL or SIGSTOP, which cannot
be caught or ignored."

Therefore, since the handler can never be executed, flag will never be set
to 1. This renders the test invalid.

Signed-off-by: Ricardo B. Marlière <rbm@suse.com>
Reviewed-by: Li Wang <liwang@redhat.com>