Rewrite openposix clock_settime bugged tests inside LTP
Both clock_settime_7-1 and clock_settime_8-1 are affected by the
same bug which is probably related to the way tests are written.

sleep() is used by the parent and is not reliable. It can oversleep
or undersleep according to system overload or signals which are
received. And we never check for its return value. Using
clock_nanosleep would make parent more reliable in this case.

At the same time, the test is taking for granted a certain
synchronization between child and parent, which is not always true
in case of system overload.

My suggestion is to rewrite the test using LTP, which has better
timing handling (see tst_timer.h).

To: ltp@lists.linux.it
Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>

---
Changes in v5:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v4: https://lore.kernel.org/r/20250717-clock_nanosleep05-v4-0-78bc62df960f@suse.com

Changes in v4:
- change description to include REALTIME instead of MONOTONIC
- Link to v3: https://lore.kernel.org/r/20250717-clock_nanosleep05-v3-0-6934fc1a2505@suse.com

Changes in v3:
- remove checkpoints
- sleep with REALTIME only, but take time in MONOTONIC
- use only clock_settime variant
- Link to v2: https://lore.kernel.org/r/20250630-clock_nanosleep05-v2-0-15522d5551c6@suse.com

Changes in v2:
- add SAFE_CLOCK_NANOSLEEP
- move child to MONOTONIC clock
- simplify child_nanosleep by passing flags
- calculate delta by checking sleep() time between beginning and ending time
- keep variants only for child
- Link to v1: https://lore.kernel.org/r/20250627-clock_nanosleep05-v1-1-1357109a2c81@suse.com

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 5,
    "change-id": "20250620-clock_nanosleep05-60b3cfba52fe",
    "prefixes": [],
    "history": {
      "v1": [
        "20250627-clock_nanosleep05-v1-1-1357109a2c81@suse.com"
      ],
      "v2": [
        "20250630-clock_nanosleep05-v2-0-15522d5551c6@suse.com"
      ],
      "v3": [
        "20250717-clock_nanosleep05-v3-0-6934fc1a2505@suse.com"
      ],
      "v4": [
        "20250717-clock_nanosleep05-v4-0-78bc62df960f@suse.com"
      ]
    }
  }
}